src/GamifyPy/backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
fastapi-mail==1.5.0
sqlalchemy==2.0.40
pydantic==2.11.5
passlib==1.7.4
argon2-cffi==25.1.0
python-dotenv==1.1.0
openai==1.95.1
httpx==0.28.1
google-auth==2.38.0
python-jose==3.4.0
python-multipart==0.0.20
starlette==0.46.2
alembic==1.16.1
PyJWT==2.10.1
requests==2.32.4
psycopg2-binary==2.9.10
